% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_tag.R
\name{add_tag}
\alias{add_tag}
\title{Add a tag to the current project}
\usage{
add_tag(tag)
}
\arguments{
\item{tag}{Character string giving the tag name.}
}
\description{
Creates a \code{.txt} file in the project's \verb{tags/} folder.
}
