% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-project.R
\name{build_project}
\alias{build_project}
\title{Build a new project from the cubicle template}
\usage{
build_project(
  name,
  path,
  root = NULL,
  use_name = FALSE,
  load_project = FALSE,
  append = NULL
)
}
\arguments{
\item{name}{Name of the project folder.}

\item{path}{Path where the project folder should be created.}

\item{root}{Optional root directory for relative paths.}

\item{use_name}{Logical. If \code{TRUE}, rename the \code{.Rproj} file by \code{name}.
If \code{FALSE} (default), keep the template \code{.Rproj} file name.}

\item{load_project}{Logical. If \code{TRUE}, open the new project in RStudio,
if available.}

\item{append}{Optional string to append to notes and project files}
}
\description{
Build a new project from the cubicle template
}
