% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_project.R
\name{delete_project}
\alias{delete_project}
\title{Remove a project from the cubicle registry}
\usage{
delete_project(project)
}
\arguments{
\item{project}{Project name/path to registered project.}
}
\description{
Remove a project from the cubicle registry
}
