% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_project.R
\name{edit_project}
\alias{edit_project}
\title{Edit metadata for a registered project}
\usage{
edit_project(project, group = NULL, timestamp = NULL)
}
\arguments{
\item{project}{Project name/path to registered project.}

\item{group}{(Optional) New group value.}

\item{timestamp}{(Optional) New time stamp value.}
}
\description{
Edit metadata for a registered project
}
