% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_project.R
\name{open_project}
\alias{open_project}
\title{Open a registered project}
\usage{
open_project(project)
}
\arguments{
\item{project}{Project name/path to registered project.}
}
\description{
Open a registered project
}
