% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_info.R
\name{project_info}
\alias{project_info}
\title{Show information for a registered project}
\usage{
project_info(project)
}
\arguments{
\item{project}{Project name/path to registered project.}
}
\description{
Show information for a registered project
}
