% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_project.R
\name{register_project}
\alias{register_project}
\title{Register an existing project in cubicle}
\usage{
register_project(path = NULL, group = NULL, timestamp = NULL)
}
\arguments{
\item{path}{Path to a project directory or \code{.Rproj} file. If \code{NULL},
the active/open project is used.}

\item{group}{(Optional) Group label to save with the project.}

\item{timestamp}{(Optional) Date of project registration. Defaults to
\code{Sys.Date()}.}
}
\description{
Register an existing project in cubicle
}
