% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/root.R
\name{reset_root}
\alias{reset_root}
\title{Reset the cubicle root}
\usage{
reset_root(save = TRUE)
}
\arguments{
\item{save}{Logical. If \code{TRUE}, revert to package default.}
}
\description{
Reset the cubicle root
}
