% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{reset_template}
\alias{reset_template}
\title{Reset to the default cubicle template}
\usage{
reset_template(save = TRUE)
}
\arguments{
\item{save}{Logical. If \code{TRUE}, revert to package default.}
}
\description{
Reset to the default cubicle template
}
