% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/root.R
\name{set_root}
\alias{set_root}
\title{Save new cubicle root as default}
\usage{
set_root(path, save = TRUE)
}
\arguments{
\item{path}{Path to the default parent directory for projects}

\item{save}{Logical. If \code{TRUE}, save this setting across R sessions.}
}
\description{
Save new cubicle root as default
}
