% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{set_template}
\alias{set_template}
\title{Set a custom cubicle template}
\usage{
set_template(path, save = TRUE)
}
\arguments{
\item{path}{Path to a folder containing a template structure}

\item{save}{Logical. If \code{TRUE}, save this setting across R sessions.}
}
\description{
Set a custom cubicle template
}
