% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_projects.R
\name{show_projects}
\alias{show_projects}
\title{Show registered cubicle projects}
\usage{
show_projects(
  levels = 1,
  arrange = "timestamp",
  group = NULL,
  timestamp = NULL,
  tags = FALSE
)
}
\arguments{
\item{levels}{Number of directory levels to print.}

\item{arrange}{Character vector of fields to sort by. Allowed values
are \code{"timestamp"}, \code{"label"}, and \code{"group"}.}

\item{group}{(Optional) Group used to filter displayed projects.}

\item{timestamp}{(Optional) Date used to filter displayed projects.}

\item{tags}{Logical. If \code{TRUE}, print tags under each project.}
}
\description{
Show registered cubicle projects
}
