% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_project.R
\name{build_project}
\alias{build_project}
\title{Build a new project from the cubicle template}
\usage{
build_project(
  name,
  path,
  root = NULL,
  group = NULL,
  register = TRUE,
  load_project = FALSE,
  use_name = FALSE,
  append = NULL
)
}
\arguments{
\item{name}{Name of the project folder.}

\item{path}{Path where the project folder should be created.}

\item{root}{(Optional) Root directory for relative paths.}

\item{group}{(Optional) Group label to save with the project when
\code{register = TRUE}.}

\item{register}{Logical. If \code{TRUE} (default), save the project in the
internal cubicle project registry.}

\item{load_project}{Logical. If \code{TRUE}, open the new project in RStudio,
if available.}

\item{use_name}{Logical. If \code{TRUE}, rename the \code{.Rproj} file by \code{name}.
If \code{FALSE} (default), keep the template \code{.Rproj} file name.}

\item{append}{(Optional) String to append to notes and project files.}
}
\description{
Build a new project from the cubicle template
}
